Persistence layer added, fix bussines layer: city pass models->services, added cityModel
