+++ weather_api/serializers/citySerializer.py
@@ -1,4 +1,4 @@
-from weather_api.models.city import City
+from weather_api.services.city import City
 
 class CitySerializer():
    
@@ -14,5 +14,26 @@ class CitySerializer():
                 'sunset' : f"{city.sunset}",
                 'geo_coordinates' : f"[{city.coord['lat']:.2f}, {city.coord['lon']:.2f}]",
                 'requested_time' : f"{city.requested_time}",
+        }
+        return city_serialized
+
+
+class CityFullSerializer():
+   
+    def serialize(self, city):
+        city_serialized = {
+                'coord' : f"{city.coord}",
+                'weather' : f"{city.weather}",
+                'base' : f"{city.base}",
+                'main' : f"{city.main}",
+                'visibility' : f"{city.visibility}",
+                'wind' : f"{city.wind}",
+                'clouds' : f"{city.clouds}",
+                'dt' : f"{city.dt}",
+                'sys' : f"{city.sys}",
+                'timezone' : f"{city.timezone}",
+                'id_city' : f"{city.id}", #Change
+                'name' : f"{city.name}",
+                'cod' : f"{city.cod}",
         }
         return city_serialized
