+++ weather_api/views.py
@@ -2,7 +2,6 @@ from django.shortcuts import render
 from rest_framework import views
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from .models.city import City
 from .serializers.citySerializer import CitySerializer
 from .querysets.cityQuery import CityQuery
 # Create your views here.
